+++ core/Cargo.toml
@@ -2,40 +2,39 @@
 name = "sa_core"
 version = "0.1.0"
 edition = "2021"
-
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 
 [dependencies]
+async-trait = "0.1"
+base64 = "0.21"
+bytes = "1.3"
+chrono = { version = "0.4", features = ["serde"] }
+fake = "2.4"
+flate2 = "1.0"
+futures = "0.3"
+futures-retry-policies = "0.2"
+handlebars = "4.2"
+html2text = "0.4"
+leaky-bucket = "0.12"
+mac_address = "1.1"
+once_cell = "1.10"
+regex = "1.5"
 reqwest = { version = "0.11", features = ["multipart", "json", "cookies", "stream"] }
 reqwest-middleware = "0.2"
 reqwest-retry = "0.2"
-chrono = { version = "0.4", features = ["serde"] }
-fake = "2.4"
-rsa = "0.8"
-flate2 = "1.0"
-futures = "0.3"
-base64 = "0.21"
-handlebars = "4.2"
-serde_urlencoded = "0.7"
-tracing = "0.1"
-mac_address = "1.1"
 retry-policies = "0.1"
-html2text = "0.4"
-once_cell = "1.10"
-regex = "1.5"
+rsa = "0.8"
 sanitize-filename = "0.4"
+scopeguard = "1.1"
 scraper = "0.14"
 serde = { version = "1.0", features = ["derive"] }
-thiserror = "1.0"
 serde_json = "1.0"
-bytes = "1.3"
-scopeguard = "1.1"
-tar = "0.4"
-tokio-stream = "0.1"
-tokio = { version = "1.17", features = ["rt", "sync", "time", "macros"] }
-task-local-extensions = "0.1"
-async-trait = "0.1"
-uuid = { version = "1.2", features = ["v4", "v5"] }
-leaky-bucket = "0.12"
+serde_urlencoded = "0.7"
 tap = "1.0"
-futures-retry-policies = "0.2"
+tar = "0.4"
+task-local-extensions = "0.1"
+thiserror = "1.0"
+tokio = { version = "1.17", features = ["rt", "sync", "time", "macros"] }
+tokio-stream = "0.1"
+tracing = "0.1"
+uuid = { version = "1.2", features = ["v4", "v5"] }
